+++ core/src/main/java/io/snabble/sdk/Snabble.kt
@@ -17,10 +17,13 @@ import com.google.gson.JsonObject
 import io.snabble.sdk.auth.TokenRegistry
 import io.snabble.sdk.checkin.CheckInLocationManager
 import io.snabble.sdk.checkin.CheckInManager
+import io.snabble.sdk.config.CustomProperty
+import io.snabble.sdk.config.ProjectId
 import io.snabble.sdk.customization.IsMergeable
 import io.snabble.sdk.events.Events
 import io.snabble.sdk.extensions.getPackageInfoCompat
 import io.snabble.sdk.payment.PaymentCredentialsStore
+import io.snabble.sdk.payment.data.FormPrefillData
 import io.snabble.sdk.utils.*
 import okhttp3.OkHttpClient
 import java.io.ByteArrayInputStream
@@ -312,6 +315,19 @@ object Snabble {
      * Set to take control over [ShoppingCart.Item.isMergeable] default behavior.
      */
     var isMergeable: IsMergeable? = null
+
+    /**
+     * Set to have PAYONE forms prefilled with the given data.
+     */
+    var formPrefillData: FormPrefillData? = null
+
+    /**
+     * Set [CustomProperty]'s to override the default behavior.
+     *
+     * Every [CustomProperty] has to be explicitly defined and implemented beforehand
+     * to be applicable for the given project.
+     */
+    val customProperties: MutableMap<Pair<CustomProperty, ProjectId>, Any> = mutableMapOf()
 
     /**
      * Setup the snabble SDK.