+++ ui/src/main/java/io/snabble/sdk/ui/cart/CheckoutBar.kt
@@ -28,6 +28,8 @@ import io.snabble.sdk.Snabble
 import io.snabble.sdk.Snabble.instance
 import io.snabble.sdk.checkout.Checkout
 import io.snabble.sdk.checkout.CheckoutState
+import io.snabble.sdk.config.ExternalBillingSubjectLength
+import io.snabble.sdk.config.ProjectId
 import io.snabble.sdk.extensions.getApplicationInfoCompat
 import io.snabble.sdk.ui.Keyguard
 import io.snabble.sdk.ui.R
@@ -320,7 +322,7 @@ open class CheckoutBar @JvmOverloads constructor(
                     if (entry.paymentMethod == PaymentMethod.TEGUT_EMPLOYEE_CARD) {
                         project.checkout.pay(entry.paymentMethod, entry.paymentCredentials)
                     } else if (entry.paymentMethod == PaymentMethod.EXTERNAL_BILLING) {
-                        SubjectAlertDialog(context)
+                        SubjectAlertDialog(context, maxSubjectLength = getMaxSubjectLength())
                             .addMessageClickListener { message ->
                                 entry.paymentCredentials.additionalData["subject"] = message
                                 project.checkout.pay(entry.paymentMethod, entry.paymentCredentials)
@@ -471,6 +473,18 @@ open class CheckoutBar @JvmOverloads constructor(
             }
         }
     }
+
+    private fun getMaxSubjectLength(): Int? = Snabble.checkedInProject.value
+        ?.id
+        ?.let { id ->
+            Snabble.customProperties
+                .getOrDefault(
+                    ExternalBillingSubjectLength to ProjectId(id),
+                    defaultValue = null
+                )
+                ?.toString()
+                ?.toInt()
+        }
 }
 
 fun interface CheckoutPreconditionHandler {